Braim's re-read and edit for Quick Start docs
Changed little bits and small details. Added some articles like "a", "an" or "the" and removed some as well.

We need them especially when we're talking about them specifically. For example: there is one specific source code that we'll talk about, so "You can find the* source code of the completed application here:".

We could also add them when we say:
"Navigate to the blabla.cs project"
since we have a noun "project", we're supposed to add an article.

But if we say:
"Navigate to blabla.cs" you see here we don't have a noun after it. so No need to add an article like "the", and this explains why I removed some "the's".